feat: add idempotency middleware and extract reusable ResponseBuffer (#91)
* feat: add idempotency middleware and extract reusable ResponseBuffer

  - Add idempotency middleware with configurable IdempotencyStore interface
  - Extract ResponseBuffer to internal/rwutil for shared use by cache/etag/idempotency
  - Optimize cache key generation: SHA-256 → strings.Builder with Grow (78% faster)
  - Update JWT TokenStore interface for consistency


* perf: add strings.Builder pre-allocation for HMAC and cache middleware

  Add b.Grow() calls to buildCanonicalRequest functions in hmac_signer.go
  and hmac_auth.go to reduce allocations. Improves HMAC auth by up to
  24% with 23-30% fewer allocations.


* fix: prevent duplicate headers in cache and idempotency middleware

  Both middleware were replaying cached headers on top of headers already
  set by other middleware (security headers, request IDs). Now they skip
  replaying any header key that's already present in the response.


---------

+++ config/etag.go
@@ -22,7 +22,7 @@ type ETagConfig struct {
 	Weak *bool
 
 	// MaxBufferSize is the maximum response body size to buffer for ETag generation (defaults to 1MB)
-	MaxBufferSize int
+	MaxBufferSize int64
 
 	// SkipStatusCodes contains status codes that should not have ETags generated (defaults to error status codes)
 	SkipStatusCodes map[int]struct{}
